+++ packages/jni-swig-stub/src/main/jni/realm_api_helpers.cpp
@@ -284,36 +284,30 @@ realm_t *open_realm_with_scheduler(int64_t config_ptr, jobject dispatchScheduler
     return realm_open(&config_clone);
 }
 
-jobject app_exception_from_app_error(JNIEnv* env, const realm_app_error_t* error) {
-    static JavaMethod app_exception_constructor(env,
-                                                JavaClassGlobalDef::app_exception_class(),
+jobject convert_to_jvm_app_error(JNIEnv* env, const realm_app_error_t* error) {
+    static JavaMethod app_error_constructor(env,
+                                                JavaClassGlobalDef::app_error(),
                                                 "<init>",
-                                                "(Ljava/lang/String;)V");
-
-    std::stringstream message;
-
-    // This is needed to avoid sending NULL to the "<<" operator as it will crash otherwise
-    const char* server_logs = "NO LOGS";
-    if (error->link_to_server_logs != nullptr) {
-        server_logs = error->link_to_server_logs;
-    }
-
-    message << error->message << " ["
-            << "error_category=" << error->error_category << ", "
-            << "error_code=" << error->error_code << ", "
-            << "link_to_server_logs=" << server_logs
-            << "]";
-
-    return env->NewObject(JavaClassGlobalDef::app_exception_class(),
-                                       app_exception_constructor,
-                                       to_jstring(env, message.str()));
+                                                "(IIILjava/lang/String;Ljava/lang/String;)V");
+    jint category = static_cast<jint>(error->error_category);
+    jint code = static_cast<jint>(error->error_code);
+    jint httpCode = static_cast<jint>(error->http_status_code);
+    jstring message = to_jstring(env, error->message);
+    jstring serverLogs = to_jstring(env, error->link_to_server_logs);
+    return env->NewObject(JavaClassGlobalDef::app_error(),
+                          app_error_constructor,
+                          category,
+                          code,
+                          httpCode,
+                          message,
+                          serverLogs);
 }
 
 void app_complete_void_callback(void *userdata, const realm_app_error_t *error) {
     auto env = get_env(true);
     static JavaClass java_callback_class(env, "io/realm/internal/interop/AppCallback");
     static JavaMethod java_notify_onerror(env, java_callback_class, "onError",
-                                          "(Ljava/lang/Throwable;)V");
+                                          "(Lio/realm/internal/interop/sync/AppError;)V");
     static JavaMethod java_notify_onsuccess(env, java_callback_class, "onSuccess",
                                             "(Ljava/lang/Object;)V");
     static JavaClass unit_class(env, "kotlin/Unit");
@@ -323,8 +317,8 @@ void app_complete_void_callback(void *userdata, const realm_app_error_t *error) 
         env->ExceptionDescribe();
         throw std::runtime_error("An unexpected Error was thrown from Java. See LogCat");
     } else if (error) {
-        jobject app_exception = app_exception_from_app_error(env, error);
-        env->CallVoidMethod(static_cast<jobject>(userdata), java_notify_onerror, app_exception);
+        jobject app_error = convert_to_jvm_app_error(env, error);
+        env->CallVoidMethod(static_cast<jobject>(userdata), java_notify_onerror, app_error);
     } else {
         jobject unit = env->NewObject(unit_class, unit_constructor);
         env->CallVoidMethod(static_cast<jobject>(userdata), java_notify_onsuccess, unit);
@@ -335,7 +329,7 @@ void app_complete_result_callback(void* userdata, void* result, const realm_app_
     auto env = get_env(true);
     static JavaClass java_callback_class(env, "io/realm/internal/interop/AppCallback");
     static JavaMethod java_notify_onerror(env, java_callback_class, "onError",
-                                          "(Ljava/lang/Throwable;)V");
+                                          "(Lio/realm/internal/interop/sync/AppError;)V");
     static JavaMethod java_notify_onsuccess(env, java_callback_class, "onSuccess",
                                             "(Ljava/lang/Object;)V");
 
@@ -346,7 +340,7 @@ void app_complete_result_callback(void* userdata, void* result, const realm_app_
         env->ExceptionDescribe();
         throw std::runtime_error("An unexpected Error was thrown from Java. See LogCat");
     } else if (error) {
-        jobject app_exception = app_exception_from_app_error(env, error);
+        jobject app_exception = convert_to_jvm_app_error(env, error);
         env->CallVoidMethod(static_cast<jobject>(userdata), java_notify_onerror, app_exception);
     } else {
         // Remember to clone user object or else it will be invalidated right after we leave this callback
@@ -548,29 +542,23 @@ void set_log_callback(realm_sync_client_config_t* sync_client_config, jobject lo
                                               });
 }
 
-jobject convert_sync_exception(JNIEnv* jenv, const realm_sync_error_t error) {
-    static JavaMethod app_exception_constructor(jenv,
-                                                JavaClassGlobalDef::sync_exception(),
-                                                "<init>",
-                                                "(Ljava/lang/String;)V");
+jobject convert_to_jvm_sync_error(JNIEnv* jenv, const realm_sync_error_t& error) {
 
-    std::stringstream message;
-    message << error.detailed_message << " ["
-            << "error_code.category='" << error.error_code.category << "', "
-            << "error_code.value='" << error.error_code.value << "', "
-            << "error_code.message='" << error.error_code.message << "', "
-            << "is_fatal='" << std::boolalpha << error.is_fatal << std::noboolalpha << "', "
-            << "is_unrecognized_by_client='" << std::boolalpha << error.is_unrecognized_by_client << std::noboolalpha << "'"
-            << "]";
+    static JavaMethod sync_error_constructor(jenv,
+                                             JavaClassGlobalDef::sync_error(),
+                                             "<init>",
+                                             "(IILjava/lang/String;Ljava/lang/String;ZZ)V");
 
-    static JavaMethod pointer_wrapper_constructor(jenv,
-                                                  JavaClassGlobalDef::sync_exception(),
-                                                  "<init>",
-                                                  "(Ljava/lang/String;)V");
+    jint category = static_cast<jint>(error.error_code.category);
+    jint value = error.error_code.value;
+    jstring msg = to_jstring(jenv, error.error_code.message);
+    jstring detailed_msg = to_jstring(jenv, error.detailed_message);
+    jboolean is_fatal = error.is_fatal;
+    jboolean is_unrecognized_by_client = error.is_unrecognized_by_client;
 
-    return jenv->NewObject(JavaClassGlobalDef::sync_exception(),
-                           pointer_wrapper_constructor,
-                           to_jstring(jenv, message.str()));
+    return jenv->NewObject(JavaClassGlobalDef::sync_error(),
+                           sync_error_constructor,
+                           category, value, msg, detailed_msg, is_fatal, is_unrecognized_by_client);
 }
 
 void sync_set_error_handler(realm_sync_config_t* sync_config, jobject error_handler) {
@@ -580,16 +568,16 @@ void sync_set_error_handler(realm_sync_config_t* sync_config, jobject error_hand
                                             auto sync_error_callback = static_cast<jobject>(userdata);
 
                                             jobject session_pointer_wrapper = wrap_pointer(jenv,reinterpret_cast<jlong>(session));
-                                            jobject sync_exception = convert_sync_exception(jenv, error);
+                                            jobject sync_error = convert_to_jvm_sync_error(jenv, error);
 
                                             static JavaMethod sync_error_method(jenv,
                                                                                 JavaClassGlobalDef::sync_error_callback(),
                                                                                 "onSyncError",
-                                                                                "(Lio/realm/internal/interop/NativePointer;Lio/realm/mongodb/SyncException;)V");
+                                                                                "(Lio/realm/internal/interop/NativePointer;Lio/realm/internal/interop/sync/SyncError;)V");
                                             jenv->CallVoidMethod(sync_error_callback,
                                                                  sync_error_method,
                                                                  session_pointer_wrapper,
-                                                                 sync_exception);
+                                                                 sync_error);
                                         },
                                         static_cast<jobject>(get_env()->NewGlobalRef(error_handler)),
                                         [](void *userdata) {